#2375 Fix Back Navigation in custom apps - move common functionality down - use same NavHostFragment - align packages in core
