+++ app/src/main/java/org/kiwix/kiwixmobile/di/modules/KiwixModule.kt
@@ -28,7 +28,7 @@ import org.kiwix.kiwixmobile.di.KiwixScope
 import org.kiwix.kiwixmobile.zim_manager.Fat32Checker
 import org.kiwix.kiwixmobile.zim_manager.FileWritingFileSystemChecker
 import org.kiwix.kiwixmobile.zim_manager.MountFileSystemChecker
-import org.kiwix.kiwixmobile.zim_manager.MountPointProducer
+import org.kiwix.kiwixmobile.core.zim_manager.MountPointProducer
 
 @Module
 object KiwixModule {
