+++ app/src/main/java/org/kiwix/kiwixmobile/nav/destination/reader/KiwixReaderFragment.kt
@@ -41,7 +41,6 @@ import org.kiwix.kiwixmobile.core.R.anim
 import org.kiwix.kiwixmobile.core.base.BaseActivity
 import org.kiwix.kiwixmobile.core.base.FragmentActivityExtensions.Super
 import org.kiwix.kiwixmobile.core.base.FragmentActivityExtensions.Super.ShouldCall
-import org.kiwix.kiwixmobile.core.base.FragmentActivityExtensions.Super.ShouldNotCall
 import org.kiwix.kiwixmobile.core.extensions.ActivityExtensions.setupDrawerToggle
 import org.kiwix.kiwixmobile.core.extensions.getAttribute
 import org.kiwix.kiwixmobile.core.extensions.setImageDrawableCompat
@@ -204,16 +203,6 @@ class KiwixReaderFragment : CoreReaderFragment() {
     if (zimReaderContainer.zimFile == null) {
       exitBook()
     }
-  }
-
-  override fun onBackPressed(activity: AppCompatActivity): Super {
-    val callType = super.onBackPressed(activity)
-    if (callType == ShouldCall && getCurrentWebView().canGoBack()) {
-      getCurrentWebView().goBack()
-    } else if (callType == ShouldCall) {
-      getActivity()?.finish()
-    }
-    return ShouldNotCall
   }
 
   override fun restoreViewStateOnInvalidJSON() {
