+++ app/src/main/res/layout/activity_kiwix_main.xml
@@ -30,7 +30,7 @@
 
     <fragment
       android:id="@+id/nav_host_fragment"
-      android:name="org.kiwix.kiwixmobile.main.NavigationHostFragment"
+      android:name="org.kiwix.kiwixmobile.core.main.NavigationHostFragment"
       android:layout_width="match_parent"
       android:layout_height="match_parent"
       app:defaultNavHost="true"
